// src/app/shared/services/camunda-api/model/taskDto.ts
//
//   Copyright 2026 Deutsches Zentrum für Luft- und Raumfahrt e.V.
//
//   Licensed under the Apache License, Version 2.0 (the "License");
//   you may not use this file except in compliance with the License.
//   You may obtain a copy of the License at
//
//       http://www.apache.org/licenses/LICENSE-2.0
//
//   Unless required by applicable law or agreed to in writing, software
//   distributed under the License is distributed on an "AS IS" BASIS,
//   WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
//   See the License for the specific language governing permissions and
//   limitations under the License.
//

/**
 * Camunda Platform REST API
 * OpenApi Spec for Camunda Platform REST API.
 *
 * The version of the OpenAPI document: 7.16.0
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */
import { CamundaFormRef } from 'src/app/shared/services/camunda-api/model/camundaFormRef';


export interface TaskDto { 
    /**
     * The task id.
     */
    id?: string | null;
    /**
     * The task name.
     */
    name?: string | null;
    /**
     * The assignee\'s id.
     */
    assignee?: string | null;
    /**
     * The owner\'s id.
     */
    owner?: string | null;
    /**
     * The date the task was created on. [Default format](https://docs.camunda.org/manual/7.16/reference/rest/overview/date-format/) `yyyy-MM-dd\'T\'HH:mm:ss.SSSZ`.
     */
    created?: string | null;
    /**
     * The task\'s due date. [Default format](https://docs.camunda.org/manual/7.16/reference/rest/overview/date-format/) `yyyy-MM-dd\'T\'HH:mm:ss.SSSZ`.
     */
    due?: string | null;
    /**
     * The follow-up date for the task. [Default format](https://docs.camunda.org/manual/7.16/reference/rest/overview/date-format/) `yyyy-MM-dd\'T\'HH:mm:ss.SSSZ`.
     */
    followUp?: string | null;
    /**
     * The task\'s delegation state. Possible values are `PENDING` and `RESOLVED`.
     */
    delegationState?: TaskDto.DelegationStateEnum | null;
    /**
     * The task\'s description.
     */
    description?: string | null;
    /**
     * The id of the execution the task belongs to.
     */
    executionId?: string | null;
    /**
     * The id the parent task, if this task is a subtask.
     */
    parentTaskId?: string | null;
    /**
     * The task\'s priority.
     */
    priority?: number | null;
    /**
     * The id of the process definition the task belongs to.
     */
    processDefinitionId?: string | null;
    /**
     * The id of the process instance the task belongs to.
     */
    processInstanceId?: string | null;
    /**
     * The id of the case execution the task belongs to.
     */
    caseExecutionId?: string | null;
    /**
     * The id of the case definition the task belongs to.
     */
    caseDefinitionId?: string | null;
    /**
     * The id of the case instance the task belongs to.
     */
    caseInstanceId?: string | null;
    /**
     * The task\'s key.
     */
    taskDefinitionKey?: string | null;
    /**
     * Whether the task belongs to a process instance that is suspended.
     */
    suspended?: boolean | null;
    /**
     * If not `null`, the form key for the task.
     */
    formKey?: string | null;
    /**
     * A reference to a specific version of a Camunda Form.
     */
    camundaFormRef?: { [key: string]: CamundaFormRef; } | null;
    /**
     * If not `null`, the tenant id of the task.
     */
    tenantId?: string | null;
}
export namespace TaskDto {
    export type DelegationStateEnum = 'PENDING' | 'RESOLVED';
    export const DelegationStateEnum = {
        Pending: 'PENDING' as DelegationStateEnum,
        Resolved: 'RESOLVED' as DelegationStateEnum
    };
}